1st Update
I've trained the model 2500 images only, because of that the image is getting cropped too. So I've updated the code but i'm unable to train the model due to my laptop not having a gpu,i request anyone to please train the model with more images and test it. Thank u